Sort out Typhoon+Infrastructure stuff
