+++ Source/Configuration/DefinitionOptionConfiguration/TyphoonDefinition+Option.m
@@ -19,6 +19,7 @@
 #import "TyphoonInjectionContext.h"
 #import "TyphoonInjection.h"
 #import "TyphoonDefinition+Internal.h"
+#import "TyphoonDefinition+Infrastructure.h"
 #import "TyphoonDefinition+InstanceBuilder.h"
 
 @interface TyphoonOptionDefinition : TyphoonFactoryDefinition <TyphoonAutoInjectionConfig>
