+++ Source/Definition/Internal/TyphoonDefinition+InstanceBuilder.h
@@ -17,13 +17,6 @@
 @class TyphoonComponentFactory;
 @class TyphoonRuntimeArguments;
 
-typedef void(^TyphoonInjectionsEnumerationBlock)(id injection, id*injectionToReplace, BOOL*stop);
-
-typedef NS_OPTIONS(NSInteger, TyphoonInjectionsEnumerationOption) {
-    TyphoonInjectionsEnumerationOptionProperties = 1 << 0,
-    TyphoonInjectionsEnumerationOptionMethods = 1 << 2,
-    TyphoonInjectionsEnumerationOptionAll = TyphoonInjectionsEnumerationOptionProperties | TyphoonInjectionsEnumerationOptionMethods,
-};
 
 @interface TyphoonDefinition (InstanceBuilder)
 
@@ -31,25 +24,6 @@ typedef NS_OPTIONS(NSInteger, TyphoonInjectionsEnumerationOption) {
 
 - (void)doInjectionEventsOn:(id)instance withArgs:(TyphoonRuntimeArguments *)args factory:(TyphoonComponentFactory *)factory;
 
-
-// TODO: cleanup these
-
-- (TyphoonMethod *)beforeInjections;
-
-- (NSSet *)injectedProperties;
-
-- (NSOrderedSet *)injectedMethods;
-
-- (TyphoonMethod *)afterInjections;
-
-- (void)enumerateInjectionsOfKind:(Class)injectionClass options:(TyphoonInjectionsEnumerationOption)options
-                       usingBlock:(TyphoonInjectionsEnumerationBlock)block;
-
-- (BOOL)hasRuntimeArgumentInjections;
-
-- (void)addInjectedProperty:(id <TyphoonPropertyInjection>)property;
-
-- (void)addInjectedPropertyIfNotExists:(id <TyphoonPropertyInjection>)property;
 
 - (id)targetForInitializerWithFactory:(TyphoonComponentFactory *)factory args:(TyphoonRuntimeArguments *)args;
