+++ Source/Factory/Internal/TyphoonFactoryPropertyInjectionPostProcessor.m
@@ -12,7 +12,7 @@
 #import "TyphoonFactoryPropertyInjectionPostProcessor.h"
 #import "TyphoonComponentFactory.h"
 #import "TyphoonDefinition.h"
-#import "TyphoonDefinition+InstanceBuilder.h"
+#import "TyphoonDefinition+Infrastructure.h"
 #import "TyphoonInjectionByComponentFactory.h"
 #import "TyphoonInjectionByType.h"
 #import "TyphoonIntrospectionUtils.h"
