+++ Source/Typhoon.h
@@ -23,28 +23,24 @@ FOUNDATION_EXPORT const unsigned char TyphoonVersionString[];
 #import "TyphoonAssembly.h"
 #import "TyphoonDefinition.h"
 #import "TyphoonBlockDefinition.h"
-#import "TyphoonFactoryDefinition.h"
-#import "TyphoonDefinition+Infrastructure.h"
 #import "TyphoonMethod.h"
 #import "TyphoonConfigPostProcessor.h"
 #import "TyphoonResource.h"
 #import "TyphoonBundleResource.h"
 #import "TyphoonComponentFactory.h"
-#import "TyphoonComponentFactory+InstanceBuilder.h"
+#import "TyphoonBlockComponentFactory.h"
 #import "TyphoonDefinitionPostProcessor.h"
 #import "TyphoonInstancePostProcessor.h"
+#import "TyphoonTypeConverter.h"
 #import "TyphoonIntrospectionUtils.h"
 #import "TyphoonCollaboratingAssemblyProxy.h"
 #import "NSObject+FactoryHooks.h"
 #import "TyphoonTestUtils.h"
 #import "TyphoonPatcher.h"
-
-#import "TyphoonBlockComponentFactory.h"
-#import "TyphoonAssemblyActivator.h"
-
 #import "TyphoonAutoInjection.h"
-
 #import "TyphoonDefinitionNamespace.h"
+
+#import "TyphoonAssemblyActivator.h"
 
 #if TARGET_OS_IPHONE
 #import "TyphooniOS.h"
