+++ Tests/Definition/TyphoonDefinitionTests.m
@@ -14,9 +14,9 @@
 #import "OCLogTemplate.h"
 #import "Knight.h"
 #import "Typhoon.h"
+#import "Typhoon+Infrastructure.h"
 #import "AutoWiringKnight.h"
 #import "TyphoonDefinition+Internal.h"
-#import "TyphoonDefinition+InstanceBuilder.h"
 #import "TyphoonReferenceDefinition.h"
 #import "TyphoonInjectionByObjectInstance.h"
 #import "TyphoonInjections.h"
